+++ Auth0/SafariWebAuth.swift
@@ -150,22 +150,21 @@ class SafariWebAuth: WebAuth {
     func newSafari(_ authorizeURL: URL, callback: @escaping (Result<Credentials>) -> Void) -> (SFSafariViewController, (Result<Credentials>) -> Void) {
         let controller = SFSafariViewController(url: authorizeURL)
         controller.modalPresentationStyle = safariPresentationStyle
-        
+
         if #available(iOS 11.0, *) {
             controller.dismissButtonStyle = .cancel
         }
-        
-        let finish: (Result<Credentials>) -> Void = { [weak controller] (result: Result<Credentials>) -> Void in
-            guard let presenting = controller?.presentingViewController else {
-                return callback(Result.failure(error: WebAuthError.cannotDismissWebAuthController))
-            }
 
+        let finish: (Result<Credentials>) -> Void = { [weak controller] (result: Result<Credentials>) -> Void in
             if case .failure(let cause as WebAuthError) = result, case .userCancelled = cause {
                 DispatchQueue.main.async {
                     callback(result)
                 }
             } else {
                 DispatchQueue.main.async {
+                    guard let presenting = controller?.presentingViewController else {
+                        return callback(Result.failure(error: WebAuthError.cannotDismissWebAuthController))
+                    }
                     presenting.dismiss(animated: true) {
                         callback(result)
                     }